requirements_ai.txt
# AI股票分析系统依赖包
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
joblib>=1.2.0
matplotlib>=3.5.0
seaborn>=0.12.0
talib-binary>=0.4.25
flask>=2.2.0
flask-cors>=4.0.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
requests>=2.28.0
asyncio
aiohttp>=3.8.0